Make most mekanism tools extend the vanilla tool class
