+++ src/main/java/mekanism/tools/common/ToolUtils.java
@@ -1,11 +1,14 @@
 package mekanism.tools.common;
 
 import javax.annotation.Nonnull;
+import mekanism.common.MekanismItems;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
+import net.minecraft.init.Items;
 import net.minecraft.item.Item.ToolMaterial;
+import net.minecraft.item.ItemStack;
 
 public class ToolUtils {
 
@@ -40,5 +43,22 @@ public class ToolUtils {
 
     public static boolean canShovelHarvest(Block block) {
         return block == Blocks.SNOW_LAYER || block == Blocks.SNOW;
+    }
+
+    public static ItemStack getRepairStack(ToolMaterial material) {
+        if (material == MekanismTools.toolOBSIDIAN || material == MekanismTools.toolOBSIDIAN2) {
+            return new ItemStack(MekanismItems.Ingot, 1, 0);
+        } else if (material == MekanismTools.toolLAZULI || material == MekanismTools.toolLAZULI2) {
+            return new ItemStack(Items.DYE, 1, 4);
+        } else if (material == MekanismTools.toolOSMIUM || material == MekanismTools.toolOSMIUM2) {
+            return new ItemStack(MekanismItems.Ingot, 1, 1);
+        } else if (material == MekanismTools.toolBRONZE || material == MekanismTools.toolBRONZE2) {
+            return new ItemStack(MekanismItems.Ingot, 1, 2);
+        } else if (material == MekanismTools.toolGLOWSTONE || material == MekanismTools.toolGLOWSTONE2) {
+            return new ItemStack(MekanismItems.Ingot, 1, 3);
+        } else if (material == MekanismTools.toolSTEEL || material == MekanismTools.toolSTEEL2) {
+            return new ItemStack(MekanismItems.Ingot, 1, 4);
+        }
+        return material.getRepairItemStack();
     }
 }
