+++ src/main/java/mekanism/tools/item/ItemMekanismHoe.java
@@ -1,112 +1,23 @@
 package mekanism.tools.item;
 
-import com.google.common.collect.Multimap;
 import java.util.List;
 import javax.annotation.Nonnull;
-import mekanism.common.item.ItemMekanism;
+import mekanism.common.Mekanism;
 import mekanism.common.util.LangUtils;
 import mekanism.common.util.StackUtils;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockDirt;
-import net.minecraft.block.state.IBlockState;
+import mekanism.tools.common.ToolUtils;
 import net.minecraft.client.util.ITooltipFlag;
-import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.SharedMonsterAttributes;
-import net.minecraft.entity.ai.attributes.AttributeModifier;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
-import net.minecraft.init.SoundEvents;
-import net.minecraft.inventory.EntityEquipmentSlot;
+import net.minecraft.item.ItemHoe;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumActionResult;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class ItemMekanismHoe extends ItemMekanism {
+public class ItemMekanismHoe extends ItemHoe {
 
-    private final float speed;
-    protected ToolMaterial toolMaterial;
-
-    public ItemMekanismHoe(ToolMaterial enumtoolmaterial) {
-        super();
-        toolMaterial = enumtoolmaterial;
-        setMaxStackSize(1);
-        setMaxDamage(enumtoolmaterial.getMaxUses());
-        setCreativeTab(CreativeTabs.TOOLS);
-        speed = enumtoolmaterial.getAttackDamage() + 1.0F;
-    }
-
-    @Override
-    public boolean getIsRepairable(ItemStack stack1, ItemStack stack2) {
-        return StackUtils.equalsWildcard(ItemMekanismTool.getRepairStack(toolMaterial), stack2) || super.getIsRepairable(stack1, stack2);
-    }
-
-    @Nonnull
-    @Override
-    @SuppressWarnings("incomplete-switch")
-    public EnumActionResult onItemUse(EntityPlayer playerIn, World worldIn, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
-        ItemStack stack = playerIn.getHeldItem(hand);
-
-        if (!playerIn.canPlayerEdit(pos.offset(facing), facing, stack)) {
-            return EnumActionResult.FAIL;
-        } else {
-            int hook = net.minecraftforge.event.ForgeEventFactory.onHoeUse(stack, playerIn, worldIn, pos);
-            if (hook != 0) {
-                return hook > 0 ? EnumActionResult.SUCCESS : EnumActionResult.FAIL;
-            }
-
-            IBlockState iblockstate = worldIn.getBlockState(pos);
-            Block block = iblockstate.getBlock();
-
-            if (facing != EnumFacing.DOWN && worldIn.isAirBlock(pos.up())) {
-                if (block == Blocks.GRASS || block == Blocks.GRASS_PATH) {
-                    setBlock(stack, playerIn, worldIn, pos, Blocks.FARMLAND.getDefaultState());
-                    return EnumActionResult.SUCCESS;
-                }
-                if (block == Blocks.DIRT) {
-                    switch (iblockstate.getValue(BlockDirt.VARIANT)) {
-                        case DIRT:
-                            setBlock(stack, playerIn, worldIn, pos, Blocks.FARMLAND.getDefaultState());
-                            return EnumActionResult.SUCCESS;
-                        case COARSE_DIRT:
-                            setBlock(stack, playerIn, worldIn, pos, Blocks.DIRT.getDefaultState().withProperty(BlockDirt.VARIANT, BlockDirt.DirtType.DIRT));
-                            return EnumActionResult.SUCCESS;
-                    }
-                }
-            }
-            return EnumActionResult.PASS;
-        }
-    }
-
-    protected void setBlock(ItemStack stack, EntityPlayer player, World worldIn, BlockPos pos, IBlockState state) {
-        worldIn.playSound(player, pos, SoundEvents.ITEM_HOE_TILL, SoundCategory.BLOCKS, 1.0F, 1.0F);
-        if (!worldIn.isRemote) {
-            worldIn.setBlockState(pos, state, 11);
-            stack.damageItem(1, player);
-        }
-    }
-
-    @Nonnull
-    @Override
-    public Multimap<String, AttributeModifier> getAttributeModifiers(@Nonnull EntityEquipmentSlot equipmentSlot, ItemStack stack) {
-        Multimap<String, AttributeModifier> multimap = super.getAttributeModifiers(equipmentSlot, stack);
-        if (equipmentSlot == EntityEquipmentSlot.MAINHAND) {
-            multimap.put(SharedMonsterAttributes.ATTACK_DAMAGE.getName(), new AttributeModifier(ATTACK_DAMAGE_MODIFIER, "Weapon modifier", 0.0D, 0));
-            multimap.put(SharedMonsterAttributes.ATTACK_SPEED.getName(), new AttributeModifier(ATTACK_SPEED_MODIFIER, "Weapon modifier", this.speed - 4.0F, 0));
-        }
-        return multimap;
-    }
-
-    @Override
-    public boolean hitEntity(ItemStack stack, EntityLivingBase target, EntityLivingBase attacker) {
-        stack.damageItem(1, attacker);
-        return true;
+    public ItemMekanismHoe(ToolMaterial tool) {
+        super(tool);
+        setCreativeTab(Mekanism.tabMekanism);
     }
 
     @Override
@@ -116,8 +27,7 @@ public class ItemMekanismHoe extends ItemMekanism {
     }
 
     @Override
-    @SideOnly(Side.CLIENT)
-    public boolean isFull3D() {
-        return true;
+    public boolean getIsRepairable(ItemStack toRepair, @Nonnull ItemStack repair) {
+        return StackUtils.equalsWildcard(ToolUtils.getRepairStack(toolMaterial), repair) || super.getIsRepairable(toRepair, repair);
     }
 }
