+++ src/main/java/mekanism/tools/item/ItemMekanismShovel.java
@@ -1,55 +1,33 @@
 package mekanism.tools.item;
 
-import com.google.common.collect.Sets;
-import java.util.Set;
+import java.util.List;
 import javax.annotation.Nonnull;
+import mekanism.common.Mekanism;
+import mekanism.common.util.LangUtils;
+import mekanism.common.util.StackUtils;
 import mekanism.tools.common.ToolUtils;
-import net.minecraft.block.Block;
-import net.minecraft.block.material.Material;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
-import net.minecraft.init.SoundEvents;
+import net.minecraft.client.util.ITooltipFlag;
+import net.minecraft.item.ItemSpade;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumActionResult;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
-public class ItemMekanismShovel extends ItemMekanismTool {
+public class ItemMekanismShovel extends ItemSpade {
 
-    private static final Set<Block> EFFECTIVE_ON = Sets.newHashSet(Blocks.CLAY, Blocks.DIRT, Blocks.FARMLAND, Blocks.GRASS, Blocks.GRAVEL, Blocks.MYCELIUM, Blocks.SAND,
-          Blocks.SNOW, Blocks.SNOW_LAYER, Blocks.SOUL_SAND, Blocks.GRASS_PATH);
-
-    public ItemMekanismShovel(ToolMaterial enumtoolmaterial) {
-        super(1.5F, -3, enumtoolmaterial, EFFECTIVE_ON);
+    public ItemMekanismShovel(ToolMaterial tool) {
+        super(tool);
+        setCreativeTab(Mekanism.tabMekanism);
     }
 
     @Override
-    public boolean canHarvestBlock(@Nonnull IBlockState state, ItemStack stack) {
-        return ToolUtils.canShovelHarvest(state.getBlock());
+    @SideOnly(Side.CLIENT)
+    public void addInformation(ItemStack itemstack, World world, List<String> list, ITooltipFlag flag) {
+        list.add(LangUtils.localize("tooltip.hp") + ": " + (itemstack.getMaxDamage() - itemstack.getItemDamage()));
     }
 
-    @Nonnull
     @Override
-    public EnumActionResult onItemUse(EntityPlayer playerIn, World worldIn, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
-        ItemStack stack = playerIn.getHeldItem(hand);
-        if (!playerIn.canPlayerEdit(pos.offset(facing), facing, stack)) {
-            return EnumActionResult.FAIL;
-        }
-        IBlockState iblockstate = worldIn.getBlockState(pos);
-        Block block = iblockstate.getBlock();
-        if (facing != EnumFacing.DOWN && worldIn.getBlockState(pos.up()).getMaterial() == Material.AIR && block == Blocks.GRASS) {
-            IBlockState iblockstate1 = Blocks.GRASS_PATH.getDefaultState();
-            worldIn.playSound(playerIn, pos, SoundEvents.ITEM_SHOVEL_FLATTEN, SoundCategory.BLOCKS, 1.0F, 1.0F);
-            if (!worldIn.isRemote) {
-                worldIn.setBlockState(pos, iblockstate1, 11);
-                stack.damageItem(1, playerIn);
-            }
-            return EnumActionResult.SUCCESS;
-        }
-        return EnumActionResult.PASS;
+    public boolean getIsRepairable(ItemStack toRepair, @Nonnull ItemStack repair) {
+        return StackUtils.equalsWildcard(ToolUtils.getRepairStack(toolMaterial), repair) || super.getIsRepairable(toRepair, repair);
     }
 }
